move native to own file
