+++ src/vs/platform/simpleBrowserElements/common/nativeSimpleBrowserElementsService.ts
@@ -3,29 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { CancellationToken } from '../../../base/common/cancellation.js';
 import { ProxyChannel } from '../../../base/parts/ipc/common/ipc.js';
-import { createDecorator } from '../../instantiation/common/instantiation.js';
 import { IMainProcessService } from '../../ipc/common/mainProcessService.js';
-import { IRectangle } from '../../window/common/window.js';
-
-export const INativeSimpleBrowserElementsService = createDecorator<INativeSimpleBrowserElementsService>('nativeSimpleBrowserElementsService');
-
-export interface IElementData {
-	readonly outerHTML: string;
-	readonly computedStyle: string;
-	readonly bounds: IRectangle;
-}
-
-export interface INativeSimpleBrowserElementsService {
-
-	readonly _serviceBrand: undefined;
-
-	// Properties
-	readonly windowId: number;
-
-	getElementData(rect: IRectangle, token: CancellationToken, cancellationId?: number): Promise<IElementData | undefined>;
-}
+import { INativeSimpleBrowserElementsService } from './simpleBrowserElements.js';
 
 // @ts-ignore: interface is implemented via proxy
 export class NativeSimpleBrowserElementsService implements INativeSimpleBrowserElementsService {
