+++ src/vs/workbench/services/simpleBrowserElements/browser/simpleBrowserElementsService.ts
@@ -5,7 +5,7 @@
 
 import { CancellationToken } from '../../../../base/common/cancellation.js';
 import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
-import { IElementData } from '../../../../platform/simpleBrowserElements/common/nativeSimpleBrowserElementsService.js';
+import { IElementData } from '../../../../platform/simpleBrowserElements/common/simpleBrowserElements.js';
 import { IRectangle } from '../../../../platform/window/common/window.js';
 
 export const ISimpleBrowserElementsService = createDecorator<ISimpleBrowserElementsService>('simpleBrowserElementsService');
